tiktok_analyzer.py
import requests
from bs4 import BeautifulSoup
import pandas as pd
import logging
import time
import random
from .base_analyzer import ContentAnalyzer

logger = logging.getLogger("content_analysis.tiktok")

class TikTokAnalyzer(ContentAnalyzer):
    """
    Class for analyzing trending content on TikTok.
    """
    def __init__(self, niche, data_dir):
        """
        Initialize the TikTok content analyzer.
        
        Args:
            niche (str): The content niche to analyze (e.g., weddings, fitness, travel)
            data_dir (str): Directory to store analysis results
        """
        super().__init__(niche, data_dir)
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36',
            'Accept-Language': 'en-US,en;q=0.9',
        }
        logger.info(f"Initialized TikTokAnalyzer for niche: {niche}")
    
    def analyze_trends(self, limit=20):
        """
        Analyze trending content on TikTok for the specified niche.
        
        Args:
            limit (int): Maximum number of trending posts to analyze
        
        Returns:
            dict: Analyzed trends data
        """
        logger.info(f"Analyzing TikTok trends for niche: {self.niche} (limit: {limit})")
        
        # Since direct TikTok API access is limited, we'll simulate trend analysis
        # In a production environment, this would use the official TikTok API or TikTok Creative Center
        
        # Simulated data structure for trending content
        self.trends_data = {
            'platform': 'tiktok',
            'niche': self.niche,
            'trending_posts': self._simulate_trending_posts(limit),
            'trending_hashtags': self._simulate_trending_hashtags(),
            'trending_sounds': self._simulate_trending_sounds(),
            'content_types': {},
            'themes': {}
        }
        
        # Identify content types and themes
        self.identify_content_types()
        self.identify_themes()
        
        logger.info(f"Completed TikTok trend analysis for {self.niche}")
        return self.trends_data
    
    def identify_content_types(self):
        """
        Identify common content types in the analyzed TikTok trends.
        
        Returns:
            dict: Content types with their frequency
        """
        if not self.trends_data or 'trending_posts' not in self.trends_data:
            logger.error("No trends data available for content type identification")
            return {}
        
        # Count content types
        content_types = {}
        for post in self.trends_data['trending_posts']:
            content_type = post['content_type']
            content_types[content_type] = content_types.get(content_type, 0) + 1
        
        # Sort by frequency
        sorted_types = {k: v for k, v in sorted(content_types.items(), key=lambda item: item[1], reverse=True)}
        
        self.trends_data['content_types'] = sorted_types
        logger.info(f"Identified content types: {sorted_types}")
        return sorted_types
    
    def identify_themes(self):
        """
        Identify common themes in the analyzed TikTok trends.
        
        Returns:
            dict: Themes with their frequency
        """
        if not self.trends_data or 'trending_posts' not in self.trends_data:
            logger.error("No trends data available for theme identification")
            return {}
        
        # Count themes
        themes = {}
        for post in self.trends_data['trending_posts']:
            for theme in post['themes']:
                themes[theme] = themes.get(theme, 0) + 1
        
        # Sort by frequency
        sorted_themes = {k: v for k, v in sorted(themes.items(), key=lambda item: item[1], reverse=True)}
        
        self.trends_data['themes'] = sorted_themes
        logger.info(f"Identified themes: {sorted_themes}")
        return sorted_themes
    
    def _simulate_trending_posts(self, limit):
        """
        Simulate trending posts data for the specified niche.
        In a production environment, this would fetch real data from TikTok.
        
        Args:
            limit (int): Maximum number of posts to simulate
        
        Returns:
            list: Simulated trending posts
        """
        # Define content types and themes based on niche
        content_types = self._get_niche_content_types()
        themes = self._get_niche_themes()
        
        # Generate simulated posts
        posts = []
        for i in range(limit):
            # Select random content type and themes
            content_type = random.choice(content_types)
            post_themes = random.sample(themes, min(random.randint(1, 3), len(themes)))
            
            # Generate engagement metrics with some randomness but trending posts have high engagement
            likes = random.randint(10000, 1000000)
            comments = random.randint(500, 50000)
            shares = random.randint(1000, 100000)
            views = random.randint(50000, 5000000)
            
            post = {
                'id': f"post_{i+1}",
                'content_type': content_type,
                'themes': post_themes,
                'hashtags': [f"#{self.niche}", f"#{random.choice(post_themes)}", f"#{content_type}"],
                'sound': f"trending_sound_{random.randint(1, 10)}",
                'engagement': {
                    'likes': likes,
                    'comments': comments,
                    'shares': shares,
                    'views': views
                },
                'engagement_rate': round((likes + comments + shares) / views * 100, 2) if views > 0 else 0
            }
            posts.append(post)
        
        # Sort by engagement rate
        posts.sort(key=lambda x: x['engagement_rate'], reverse=True)
        return posts
    
    def _simulate_trending_hashtags(self):
        """
        Simulate trending hashtags for the specified niche.
        
        Returns:
            list: Simulated trending hashtags
        """
        # Base hashtags on niche
        base_hashtags = [
            f"#{self.niche}",
            f"#{self.niche}tips",
            f"#{self.niche}ideas",
            f"#{self.niche}inspiration",
            f"#{self.niche}trends"
        ]
        
        # Add niche-specific hashtags
        niche_hashtags = self._get_niche_hashtags()
        
        # Combine and return
        trending_hashtags = base_hashtags + niche_hashtags
        return trending_hashtags
    
    def _simulate_trending_sounds(self):
        """
        Simulate trending sounds for the specified niche.
        
        Returns:
            list: Simulated trending sounds
        """
        # Generic trending sounds
        sounds = [
            {"name": "Original Sound", "id": "sound_1", "popularity": 95},
            {"name": "Trending Song 1", "id": "sound_2", "popularity": 90},
            {"name": "Viral Audio Clip", "id": "sound_3", "popularity": 85},
            {"name": "Popular Remix", "id": "sound_4", "popularity": 80},
            {"name": "Trending Sound Effect", "id": "sound_5", "popularity": 75}
        ]
        return sounds
    
    def _get_niche_content_types(self):
        """
        Get content types relevant to the specified niche.
        
        Returns:
            list: Niche-specific content types
        """
        # Common content types across niches
        common_types = ["tutorial", "tips", "showcase", "transformation", "review"]
        
        # Niche-specific content types
        niche_types = {
            "weddings": ["venue_tour", "dress_reveal", "proposal_story", "wedding_dance", "diy_decor"],
            "fitness": ["workout_routine", "before_after", "meal_prep", "form_check", "challenge"],
            "travel": ["destination_guide", "packing_tips", "hidden_gem", "travel_hack", "scenic_view"],
            "food": ["recipe", "restaurant_review", "cooking_hack", "taste_test", "food_asmr"],
            "beauty": ["makeup_tutorial", "skincare_routine", "product_review", "transformation", "hack"],
            "fashion": ["outfit_ideas", "styling_tips", "haul", "trend_alert", "thrift_flip"],
            "tech": ["unboxing", "review", "comparison", "tutorial", "tech_news"],
            "gaming": ["gameplay", "review", "tips_tricks", "reaction", "montage"],
            "education": ["explainer", "fact", "how_to", "study_tips", "book_review"],
            "business": ["business_tip", "success_story", "productivity_hack", "side_hustle", "interview"]
        }
        
        # Get types for the specified niche, or use common types if niche not found
        return niche_types.get(self.niche.lower(), common_types)
    
    def _get_niche_themes(self):
        """
        Get themes relevant to the specified niche.
        
        Returns:
            list: Niche-specific themes
        """
        # Niche-specific themes
        niche_themes = {
            "weddings": ["romantic", "elegant", "budget_friendly", "diy", "traditional", "modern", "seasonal"],
            "fitness": ["strength", "cardio", "flexibility", "nutrition", "motivation", "recovery", "beginner"],
            "travel": ["adventure", "luxury", "budget", "solo", "family", "cultural", "foodie", "photography"],
            "food": ["healthy", "comfort", "quick", "gourmet", "vegan", "dessert", "breakfast", "dinner"],
            "beauty": ["natural", "glam", "everyday", "special_occasion", "affordable", "luxury", "seasonal"],
            "fashion": ["casual", "formal", "streetwear", "vintage", "minimalist", "seasonal", "sustainable"],
            "tech": ["innovation", "comparison", "budget", "premium", "productivity", "entertainment", "smart_home"],
            "gaming": ["strategy", "action", "rpg", "multiplayer", "mobile", "console", "pc", "retro"],
            "education": ["study_tips", "career_advice", "language_learning", "science", "history", "math", "literature"],
            "business": ["entrepreneurship", "marketing", "finance", "leadership", "startup", "remote_work", "networking"]
        }
        
        # Get themes for the specified niche, or return generic themes if niche not found
        return niche_themes.get(self.niche.lower(), ["trending", "viral", "popular", "educational", "entertaining"])
    
    def _get_niche_hashtags(self):
        """
        Get hashtags relevant to the specified niche.
        
        Returns:
            list: Niche-specific hashtags
        """
        # Niche-specific hashtags
        niche_hashtags = {
            "weddings": ["#weddingday", "#bridetobe", "#weddingplanning", "#weddingdress", "#weddingphotography"],
            "fitness": ["#fitnessmotivation", "#workout", "#gym", "#healthylifestyle", "#fitfam"],
            "travel": ["#travelphotography", "#wanderlust", "#travelgram", "#adventure", "#exploremore"],
            "food": ["#foodie", "#foodporn", "#homemade", "#cooking", "#foodphotography"],
            "beauty": ["#makeup", "#skincare", "#beauty", "#makeuptutorial", "#glam"],
            "fashion": ["#ootd", "#style", "#fashionista", "#outfitinspo", "#fashionstyle"],
            "tech": ["#technology", "#gadgets", "#innovation", "#techreview", "#newtech"],
            "gaming": ["#gamer", "#gaming", "#videogames", "#gamingcommunity", "#esports"],
            "education": ["#learning", "#studytips", "#studentlife", "#education", "#knowledge"],
            "business": ["#entrepreneur", "#business", "#success", "#marketing", "#smallbusiness"]
        }
        
        # Get hashtags for the specified niche, or return generic hashtags if niche not found
        return niche_hashtags.get(self.niche.lower(), ["#trending", "#viral", "#foryou", "#fyp", "#popular"])
