requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
pillow==10.0.0
opencv-python==4.8.0.74
moviepy==1.0.3
flask==2.3.3
sqlalchemy==2.0.20
schedule==1.2.0

# API integrations
instagrapi==1.18.4
tiktok-api==0.1.5
unsplash-python==1.0.0
pexels-api==1.0.1
openai==1.3.0
buffer-python==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Documentation
sphinx==7.1.2
